feat(ui): overhaul Chainlit UI to Forensic premium standards
