+++ compose.yml
@@ -6,7 +6,6 @@ x-var:
 # Reusable environment for all services
 x-common-env: &common-env
   HF_TOKEN: ${HF_TOKEN:-}
-  GRADIO_ANALYTICS_ENABLED: "False"
   HF_HUB_OFFLINE: "1"
   TRANSFORMERS_OFFLINE: "1"
   EMBED_MODEL: /model
@@ -212,7 +211,7 @@ services:
     environment: *common-env
     command: /app/.venv/bin/python scripts/generate_cache_manifest.py validate
 
-  # True End-to-End tests against the live running Gradio application.
+  # True End-to-End tests against the live running Chainlit application.
   # Verifies actual HTTP responses from both Dev (Ollama) and Staging (HF).
   test-e2e:
     container_name: vexilon-test-e2e
